skills/kooui/aesthetic-scorer/requirements.txt
# Core dependencies for Aesthetic Scorer Skill
# Install with: pip install -r requirements.txt

# PyTorch and torchvision
torch>=1.8.0
torchvision>=0.9.0

# Image processing
Pillow>=8.0.0

# HTTP requests
requests>=2.25.0

# Numerical computing
numpy>=1.19.0

# Optional: for visualization
matplotlib>=3.3.0

# Optional: for progress bars
tqdm>=4.50.0
